feat(deploy): add Redis pubsub and statestore components to Kubernetes deployment using Dapr
Two new files were added to the deploy directory, `redis-pubsub.yaml` and `redis-statestore.yaml`. These files define Kubernetes components for Redis pubsub and statestore using Dapr. The `redisHost` metadata field is set to `redis-master.denis.svc.cluster.local:6379` and the `redisPassword` metadata field is set to a Kubernetes secret named `redis` with a key of `redis-password`. TLS is commented out but can be enabled if needed.