feat(deploy): add Kubernetes namespace and deployment for integration service
Add a new Kubernetes namespace named 'denis' with annotations for Linkerd service mesh and enable injection of sidecar proxies. Also, add a new deployment for the integration service with a single replica and a container running the latest version of the 'ghcr.io/shumkovdenis/bl-integration' image. The container is configured to listen on port 6001 and has environment variable PORT set to 6001. The deployment also has annotations for Dapr service mesh to enable the integration service to communicate with other Dapr-enabled services.