feat(integration): add helpers package with trace and logger interceptors and middlewares
The commit adds a new helpers package to the integration service. The package contains trace and logger interceptors and middlewares. The trace interceptor is used to extract trace information from incoming requests and add it to outgoing requests. The logger interceptor is used to log incoming requests and outgoing responses. The logger middleware is used to log incoming requests. The trace middleware is used to extract trace information from incoming requests and add it to the request context.