feat(remote): add support for tracing in HTTP and gRPC requests
Add support for tracing in HTTP and gRPC requests by adding the `TraceConfig` struct to the `helpers` package and using it in the `NewTraceInterceptor` and `NewClientTraceMiddleware` functions. The `TraceConfig` struct has three boolean fields: `UseTraceParentHeader`, `UseTraceStateHeader`, and `UseGrpcTraceBinHeader`, which are used to determine which headers to set in the request. The `Config` struct now has two new fields: `HTTPTrace` and `GRPCTrace`, which are used to configure tracing for HTTP and gRPC requests, respectively.