feat: add tracing configuration to services
Add tracing configuration to all services by adding environment variables to the deployment files. The environment variables are used to configure the tracing settings for the services. The following environment variables were added to each service:
- HTTP_TRACE_TRACE_PARENT
- HTTP_TRACE_TRACE_STATE
- HTTP_TRACE_GRPC_TRACE_BIN
- GRPC_TRACE_TRACE_PARENT
- GRPC_TRACE_TRACE_STATE
- GRPC_TRACE_GRPC_TRACE_BIN

Also, change the app-protocol to grpc in the integration-grpc-service.yaml and remote.yaml files.

Add a launch.json file to the gateway service to enable debugging in VSCode.

In the grpc.go file, set the dapr-app-id header to "remote" to enable tracing.

Update the test.http file to include a test for the new /http endpoint and a test for the /grpc endpoint.