feat: add support for retry with count
This change adds support for retrying calls with a count parameter in certain circumstances when an error occurs. A new `ConnectError` error type has been introduced for this purpose, along with a new call to `newCallError()` inside `Call()` to handle errors within the Connect service. A new `extractCallError()` function has also been added to handle messages. This change also updates the dependencies of the project by updating the go.mod file.