feat: Add Kubernetes deployment and renaming components
This commit adds `kube-apply` and  `kube-apply-services` scripts to the Makefile for easier Kubernetes deployment. It also renames some files in the `deploy` directory, reorganizing them into a `components` subdirectory. Lastly, `remote.yaml` was removed and two new YAML deployment files, `service-a.yaml` and `service-b.yaml`, were added for grpc services. This should make it easier to manage and deploy the services, especially when there are multiple remote services.