feat: Add IsBinary property to the Config struct
Add a new property to the Config struct called IsBinary, which can be set in the environment or will default to false if not set. This property is used by both the gRPC and HTTP callees to determine which tracing middleware to use. Removed the IsBinary() method since it's no longer needed.